Move mapped attributes to a separate tab in the local claims edit page
