+++ apps/admin-portal/src/components/claims/edit/edit-basic-details-local-claims.tsx
@@ -30,182 +30,29 @@ export const EditBasicDetailsLocalClaims = (
     props: EditBasicDetailsLocalClaimsPropsInterface
 ): React.ReactElement => {
 
-    const [mappedAttributes, setMappedAttributes] = useState<Set<number>>(new Set([0]));
-    const [userStore, setUserStore] = useState([]);
     const [claimURIBase, setClaimURIBase] = useState("");
 
     const { claim, update } = props;
 
     useEffect(() => {
-        const userstore = [];
-        userstore.push({
-            id: "PRIMARY",
-            name: "PRIMARY"
-        });
-        getUserStoreList().then((response) => {
-            userstore.push(...response.data);
-            setUserStore(userstore);
-        }).catch(error => {
-            setUserStore(userstore);
-            // TODO: Notify
-        });
-
         getADialect("local").then((response) => {
             setClaimURIBase(response.dialectURI);
         }).catch(error => {
 
         })
 
-        const tempMappedAttributes = new Set(mappedAttributes);
-        claim?.attributeMapping.forEach((attribute, index) => {
-            tempMappedAttributes.add(index);
-        });
-        setMappedAttributes(tempMappedAttributes);
-
     }, []);
-
-    const getMappedAttributes = (values: Map<string, FormValue>): AttributeMapping[] => {
-        const attributes: AttributeMapping[] = [];
-        mappedAttributes.forEach((attribute: number) => {
-            attributes.push({
-                mappedAttribute: values.get("attribute" + attribute).toString(),
-                userstore: values.get("userstore" + attribute).toString()
-            });
-        });
-        return attributes;
-    }
-
-    const generateMappedAttributes = (): React.ReactElement[] => {
-
-        const mappedElements: React.ReactElement[] = [];
-        mappedAttributes?.forEach((attribute: number) => {
-
-            const isFirstElement: boolean = attribute === Array.from(mappedAttributes)[0];
-            const isOnlyElement: boolean = mappedAttributes.size === 1;
-
-            mappedElements.push(
-                <Grid.Row key={attribute} columns={3}>
-                    <Grid.Column width={7}>
-                        <Field
-                            type="dropdown"
-                            name={"userstore" + attribute}
-                            label={isFirstElement ? "User Store" : null}
-                            required={true}
-                            requiredErrorMessage="Select a user store"
-                            placeholder="Select a user store"
-                            value={claim?.attributeMapping[attribute]?.userstore}
-                            children={
-                                userStore.map(store => {
-                                    return {
-                                        key: store.id,
-                                        value: store.id,
-                                        text: store.name
-                                    }
-                                })
-                            }
-                            validation={
-                                (
-                                    value: string,
-                                    validation: Validation,
-                                    values: Map<string, FormValue>
-                                ) => {
-                                    let isSameUserStore = false;
-                                    let mappedAttribute;
-                                    for (mappedAttribute of mappedAttributes) {
-                                        if (
-                                            (values.get("userstore" + mappedAttribute)
-                                                === value)
-                                            && mappedAttribute !== attribute
-                                        ) {
-                                            isSameUserStore = true;
-                                            break;
-                                        }
-                                    };
-                                    if (isSameUserStore) {
-                                        validation.isValid = false;
-                                        validation.errorMessages.push(
-                                            "This User Store has been selected twice. " +
-                                            "A User Store can only be selected once."
-                                        )
-                                    }
-                                }
-                            }
-                            displayErrorOn="blur"
-                        />
-                    </Grid.Column>
-                    <Grid.Column width={7}>
-                        <Field
-                            type="text"
-                            name={"attribute" + attribute}
-                            label={isFirstElement ? "Attribute to map to" : null}
-                            required={true}
-                            requiredErrorMessage="Enter an attribute or delete the mapping"
-                            placeholder="Enter an attribute"
-                            value={claim?.attributeMapping[attribute]?.mappedAttribute}
-                        />
-                    </Grid.Column>
-                    <Grid.Column width={2} verticalAlign="bottom">
-                        {
-                            !isOnlyElement
-                                ? (
-                                    <Button
-                                        type="button"
-                                        size="mini"
-                                        primary
-                                        circular
-                                        icon={"trash"}
-                                        onClick={() => {
-                                            const tempMappedAttributes = new Set(mappedAttributes);
-                                            if (!isOnlyElement) {
-                                                tempMappedAttributes.delete(attribute);
-                                            }
-                                            setMappedAttributes(tempMappedAttributes);
-                                        }}
-                                    />
-                                )
-                                : null
-                        }
-                    </Grid.Column>
-                </Grid.Row>
-            )
-        });
-
-        const lastElement: number = Array.from(mappedAttributes)[mappedAttributes.size - 1];
-        mappedAttributes.size < userStore.length
-            ? mappedElements.push(
-                <Grid.Row key={lastElement + 1} textAlign="center" columns={1}>
-                    <Grid.Column width={16}>
-                        <Button
-                            type="button"
-                            size="mini"
-                            primary
-                            circular
-                            icon="add"
-                            onClick={() => {
-                                if (mappedAttributes.size < userStore.length) {
-                                    const tempMappedAttributes = new Set(mappedAttributes);
-                                    tempMappedAttributes.add(lastElement + 1);
-                                    setMappedAttributes(tempMappedAttributes);
-                                }
-                            }}
-                        />
-                    </Grid.Column>
-                </Grid.Row>
-            )
-            : null;
-        return mappedElements;
-    };
 
     return (
         <Forms
             onSubmit={(values) => {
-                const data:Claim = {
+                const data: Claim = {
                     claimURI: claimURIBase + "/" + values.get("claimURI").toString(),
                     description: values.get("description").toString(),
                     displayOrder: parseInt(values.get("displayOrder").toString()),
                     regEx: values.get("regularExpression").toString(),
                     displayName: values.get("name").toString(),
-                    attributeMapping: getMappedAttributes(values),
+                    attributeMapping: claim.attributeMapping,
                     properties: claim.properties,
                     supportedByDefault: values.get("supportedByDefault").length > 0,
                     readOnly: values.get("readOnly").length > 0,
@@ -248,7 +95,7 @@ export const EditBasicDetailsLocalClaims = (
                             required={true}
                             requiredErrorMessage="Claim URI is required"
                             placeholder="Enter a claim URI"
-                            value={claim?.claimURI.replace(claimURIBase+"/","")}
+                            value={claim?.claimURI.replace(claimURIBase + "/", "")}
                         />
                         <Field
                             type="text"
@@ -277,7 +124,7 @@ export const EditBasicDetailsLocalClaims = (
                             required={false}
                             requiredErrorMessage=""
                             children={[{ value: "Support", label: "" }]}
-                            value={claim?.supportedByDefault?["Support"]:[]}
+                            value={claim?.supportedByDefault ? ["Support"] : []}
                         />
                         <Field
                             type="checkbox"
@@ -303,16 +150,6 @@ export const EditBasicDetailsLocalClaims = (
                 </Grid.Row>
                 <Grid.Row columns={1}>
                     <Grid.Column width={6}>
-                        <h5>Map Attributes</h5>
-                        <Grid>
-                            {
-                                generateMappedAttributes()
-                            }
-                        </Grid>
-                    </Grid.Column>
-                </Grid.Row>
-                <Grid.Row columns={1}>
-                    <Grid.Column width={6}>
                         <Field
                             type="submit"
                             value="Update"
@@ -322,4 +159,4 @@ export const EditBasicDetailsLocalClaims = (
             </Grid>
         </Forms>
     )
-}
+};
