+++ apps/admin-portal/src/components/claims/edit/index.ts
@@ -18,3 +18,4 @@
 
 export * from "./edit-additional-properties-local-claims";
 export * from "./edit-basic-details-local-claims";
+export * from "./edit-mapped-attributes-local-claims";
